catalog-model: move EntityRelationSpec to catalog-backend
