+++ .changeset/lovely-chairs-confess.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend': patch
+---
+
+Added `EntityRelationSpec`, which was moved over from `@backstage/catalog-model`.
