+++ .changeset/lovely-falcons-hang.md
@@ -1,0 +1,5 @@
+---
+'@backstage/catalog-model': minor
+---
+
+**BREAKING**: Removed `EntityRelationSpec` as it is only used during the catalog during the catalog processing.
